pytestplugin: move pytest hooks into hook.py
Move the pytest configure hook into a generic hook file to seperate it from the
reporter implementation.
