+++ labgrid/pytestplugin/hooks.py
@@ -1,0 +1,42 @@
+import logging
+import os
+import pytest
+
+from .. import Environment
+from ..consoleloggingreporter import ConsoleLoggingReporter
+from .reporter import StepReporter
+
+@pytest.hookimpl(trylast=True)
+def pytest_configure(config):
+    terminalreporter = config.pluginmanager.getplugin('terminalreporter')
+    capturemanager = config.pluginmanager.getplugin('capturemanager')
+    rewrite = True
+    lg_log = config.option.lg_log
+    if capturemanager._method == "no":
+        rewrite = False  # other output would interfere with our rewrites
+    if terminalreporter.verbosity > 1:  # enable with -vv
+        config.pluginmanager.register(StepReporter(terminalreporter, rewrite=rewrite))
+    if terminalreporter.verbosity > 2:  # enable with -vvv
+        logging.getLogger().setLevel(logging.DEBUG)
+    if lg_log:
+        ConsoleLoggingReporter(lg_log)
+    env_config = config.option.env_config
+    lg_env = config.option.lg_env
+    lg_coordinator = config.option.lg_coordinator
+
+    if lg_env is None:
+        if env_config is not None:
+            config.warn(
+                'LG-C1',
+                "deprecated option --env-config (use --lg-env instead)",
+                __file__)
+            lg_env = env_config
+
+    env = None
+    if lg_env is None:
+        lg_env = os.environ.get('LG_ENV')
+    if lg_env is not None:
+        env = Environment(config_file=lg_env)
+    if lg_coordinator is not None:
+        env.config.set_option('crossbar_url', lg_coordinator)
+    config._labgrid_env = env
